Add Watch of Velero Backup and remove Backup controller
Change which removes Backup Controller and adds the Watch of the
Backup object within NonAdminBackup Controller.

Added two main predicates with 'upper' logic within composite
predicate to easily distinguish between predicate for NonAdminBackup
and the Backup one.

The reconcile loop updates the Status and the Spec of the NonAdminBackup
from the Velero Backup object. The Spec in the Velero Backup is updated
by the Velero, so we use similar logic to update our NonAdminBackup one.

Signed-off-by: Michal Pryc <mpryc@redhat.com>