Rework deletion of Non Admin Backup object
In this impmenetaion the finalizer is set before velero backup
is created. Once the user initializes deletion of NAB by setting
spec field then the finalizer is uplifted when the velero object
gets removed or it's not existing.

If the user initializes deletion of NAB by invoking delete api call
the velero object is not automatically removed, so until it's
manually removed or non-existing the NAB object will not get deleted
due to existing finalizer on NAB.

Signed-off-by: Michal Pryc <mpryc@redhat.com>