Fix "cannot range over" local lint errors
With the #122 local error on the make lint
appeared. This is not visible in the CI, however local go version
and lint version are the same, so may be associated with other
dependent libraries.

Still this is an issue when submitting a PR from the system where
this linting error is visible.

For error message and insights:
  https://github.com/migtools/oadp-non-admin/pull/122#issuecomment-2515304578

Signed-off-by: Michal Pryc <mpryc@redhat.com>